requirements.txt
torch
torchvision
pillow
matplotlib
pandas
streamlit
pytorch-grad-cam
dice-ml
scikit-learn
